requirements.txt.txt
streamlit==1.28.0
neo4j==5.15.0
zhipuai==2.0.1
streamlit-authenticator==0.2.3
PyYAML==6.0.1
